move envoy utils
